safari-600.7-branch/LayoutTests/http/tests/websocket/tests/hybi/too-long-payload-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/too-long-payload' failed: WebSocket frame length too large: 9223372036854775808 bytes
Tests whether WebSocket correctly aborts the connection when it receives a frame with too long payload.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
